requirements.txt
selenium>=4.1.0
pandas>=1.3.0
PyYAML>=6.0
